fix transaction table in db

+++ migrations/2020-05-04-182438_crate_transactions/up.sql
@@ -2,11 +2,11 @@ CREATE TABLE transactions(
     id INTEGER NOT NULL PRIMARY KEY,
     user_id INTEGER NOT NULL,
     date_transaction TEXT NOT NULL,
-    sell_amount INTEGER NOT NULL,
+    sell_amount REAL NOT NULL,
     sell_currency TEXT NOT NULL,
-    buy_amount INTEGER NOT NULL,
+    buy_amount REAL NOT NULL,
     buy_currency TEXT NOT NULL,
-    price_for_one INTEGER NOT NULL,
+    price_for_one REAL NOT NULL,
     FOREIGN KEY (user_id)
        REFERENCES users (id)
 );
