+++ src/models.rs
@@ -38,11 +38,11 @@ pub struct Transaction {
     pub id: i32,
     pub user_id: i32,
     pub date_transaction: String,
-    pub sell_amount: i32,
+    pub sell_amount: f32,
     pub sell_currency: String,
-    pub buy_amount: i32,
+    pub buy_amount: f32,
     pub buy_currency: String,
-    pub price_for_one: i32,
+    pub price_for_one: f32,
 }
 
 #[derive(Insertable)]
@@ -50,9 +50,9 @@ pub struct Transaction {
 pub struct NewTransaction {
     pub user_id: i32,
     pub date_transaction: String,
-    pub sell_amount: i32,
+    pub sell_amount: f32,
     pub sell_currency: String,
-    pub buy_amount: i32,
+    pub buy_amount: f32,
     pub buy_currency: String,
-    pub price_for_one: i32,
+    pub price_for_one: f32,
 }
